Call `keyUp`/`dialUp` for short presses
If the action does't need to know about key up events that were long presses they can use to old method, which is only called on short presses.

This is useful for actions that perform one action on a long press (while the key is held) and another for short presses.

If an action only wants perform its action when the key is released, and choose what to do based on the press length, implement `keyUp(device:payload:longPress:)`.