Change `PluginCommunication` to an actor
Users had reported issues with a plugin and I was able to capture some crashes while debugging related to action instances not existing. My guess is this is related to changing the WebSockets code to be async, without handling possible data races.

I think moving `PluginCommunication` to an actor will isolate the issue ensuring messages and actions match up